Refactored code. Moved logic that is not related to VIVO in Vitro.
