+++ api/src/main/java/edu/cornell/mannlib/vitro/webapp/controller/api/dto/AuthorDTO.java
@@ -1,4 +1,4 @@
-package edu.cornell.mannlib.vitro.webapp.controller.software;
+package edu.cornell.mannlib.vitro.webapp.controller.api.dto;
 
 public class AuthorDTO {
